task 9 completed

+++ models/articles.model.js
@@ -105,9 +105,19 @@ exports.dbPatchArticle = (article_id, inc_votes) => {
 
 
 exports.dbDeleteComment = (comment_id) => {
-    return  db.query(
-        `DELETE FROM comments
-        WHERE comment_id = $1`,
+    return db.query(
+    `SELECT * FROM comments WHERE comment_id = $1`,
         [comment_id]
-    )
-}
+
+    ).then((result) => {
+        if (result.rowCount === 0) {
+            return Promise.reject({ status: 404, msg: 'Comment not found' });
+        }
+
+        return db.query(
+         `DELETE FROM comments
+            WHERE comment_id = $1`,
+            [comment_id]
+        );
+    });
+};
